[Move] to component dir
