+++ src/base/browser/secondary/contextMenu/contextMenu.ts
@@ -1,8 +1,8 @@
 import { CONTEXT_MENU_ITEM_HEIGHT, CONTEXT_MENU_SEPERATOR_HEIGHT, CONTEXT_MENU_WIDTH, IMenuItem, IMenuItemOption, MenuItem } from "src/base/browser/secondary/contextMenu/menuItem";
 import { Dimension } from "src/base/common/util/size";
-import { IComponentService } from "src/code/browser/service/componentService";
+import { IComponentService } from "src/code/browser/service/component/componentService";
 import { IThemeService } from "src/code/browser/service/theme/themeService";
-import { Component, IComponent } from "src/code/browser/workbench/component";
+import { Component, IComponent } from "src/code/browser/service/component/component";
 
 export const enum ContextMenuType {
     actionBar,
