+++ src/code/browser/service/contextMenuService.ts
@@ -2,7 +2,7 @@ import { ActionBarContextMenu } from "src/base/browser/secondary/contextMenu/act
 import { ExplorerViewContextMenu } from "src/base/browser/secondary/contextMenu/actionView/explorerViewContextMenu";
 import { ContextMenuType, Coordinate, ContextMenuDimension, IContextMenu } from "src/base/browser/secondary/contextMenu/contextMenu";
 import { EditorContextMenu } from "src/base/browser/secondary/contextMenu/editor/editorContextMenu";
-import { IComponentService } from "src/code/browser/service/componentService";
+import { IComponentService } from "src/code/browser/service/component/componentService";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 
 export const IContextMenuService = createDecorator<IContextMenuService>('context-menu-service');
