+++ src/code/browser/workbench/workspace/titleBar/windowBar.ts
@@ -2,9 +2,9 @@ import { Icons } from "src/base/browser/icon/icons";
 import { WidgetBar } from "src/base/browser/secondary/widgetBar/widgetBar";
 import { Orientation } from "src/base/common/dom";
 import { IpcChannel } from "src/base/common/ipcChannel";
-import { IComponentService } from "src/code/browser/service/componentService";
+import { IComponentService } from "src/code/browser/service/component/componentService";
 import { IThemeService } from "src/code/browser/service/theme/themeService";
-import { Component } from "src/code/browser/workbench/component";
+import { Component } from "src/code/browser/service/component/component";
 import { TitleBarComponentType } from "src/code/browser/workbench/workspace/titleBar/titleBar";
 import { WindowButton } from "src/code/browser/workbench/workspace/titleBar/windowButton";
 import { IHostService } from "src/code/platform/host/common/hostService";
